allows operations with float numbers

+++ Calculator/Calculator.cs
@@ -7,19 +7,19 @@ using System.Threading.Tasks;
 namespace CalculatorApp
 {
     public class Calculator {
-        public int Add(int num1, int num2) {
+        public float Add(float num1, float num2) {
             return num1 + num2;
         }
 
-        public int Subtract(int num1, int num2) {
+        public float Subtract(float num1, float num2) {
             return num1 - num2;
         }
 
-        public int Multiply(int num1, int num2) {
+        public float Multiply(float num1, float num2) {
             return num1 * num2;
         }
 
-        public int? Divide(int num1, int num2) {
+        public float? Divide(float num1, float num2) {
             if (num2 == 0) {
                 Console.WriteLine("Error: Can't divide by 0.");
                 return null;
